Refurbished Citizen faction into Drifter.
As a word of advice, don't let everything run off plugins. It'll get convoluted and confusing very fast.

I've done this the schematic way now, without having to be dependant on plugins. Some things require to be filled in as for models.
