+++ cwhl2rp/schema/classes/sh_drifter.lua
@@ -5,10 +5,10 @@
 
 local CLASS = Clockwork.class:New("Drifter");
 	CLASS.color = Color(66, 134, 244, 255);
-	CLASS.factions = {FACTION_CITIZEN};
+	CLASS.factions = {FACTION_DRIFTER};
 	CLASS.isDefault = true;
 	CLASS.wagesName = "Supplies";
 	CLASS.wages = 0;
-	CLASS.description = "A regular human citizen enslaved by the Universal Union.";
+	CLASS.description = "A regular human.";
 	CLASS.defaultPhysDesc = "Wearing dirty clothes.";
-CLASS_CITIZEN = CLASS:Register();
+CLASS_DRIFTER = CLASS:Register();
